+++ test/gurka/test_expansion.cc
@@ -106,6 +106,7 @@ protected:
     bool cost = false;
     bool expansion_type = false;
     bool travel_mode = false;
+    bool expansion_index = false;
 
     const std::unordered_map<std::string, int> prop_count;
     for (const auto& prop : props) {
@@ -133,6 +134,9 @@ protected:
       if (prop == "travel_mode") {
         travel_mode = true;
       }
+      if (prop == "expansion_index") {
+        expansion_index = true;
+      }
     }
     ASSERT_EQ(parsed_api.expansion().geometries_size(), exp_feats);
     ASSERT_EQ(parsed_api.expansion().edge_status_size(), edge_status ? exp_feats : 0);
@@ -143,6 +147,7 @@ protected:
     ASSERT_EQ(parsed_api.expansion().costs_size(), cost ? exp_feats : 0);
     ASSERT_EQ(parsed_api.expansion().expansion_type_size(), expansion_type ? exp_feats : 0);
     ASSERT_EQ(parsed_api.expansion().travel_modes_size(), travel_mode ? exp_feats : 0);
+    ASSERT_EQ(parsed_api.expansion().expansion_index_size(), expansion_index ? exp_feats : 0);
   }
   void check_result_json(const std::string& action,
                          const std::vector<std::string>& waypoints,
@@ -311,7 +316,9 @@ INSTANTIATE_TEST_SUITE_P(ExpandPropsTest,
                                            std::vector<std::string>{"distance", "duration",
                                                                     "pred_edge_id", "expansion_type"},
                                            std::vector<std::string>{"edge_id", "cost"},
-                                           std::vector<std::string>{"edge_id", "travel_mode"}));
+                                           std::vector<std::string>{"edge_id", "travel_mode"},
+                                           std::vector<std::string>{"expansion_index",
+                                                                    "expansion_type"}));
 
 TEST(StandAlone, MultiModalAStarModes) {
   const std::string ascii_map = R"(